+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Added
 - Implemented :meth:`sheraf.attributes.enum.EnumAttribute.__repr__`
 - Ended support for python 3.6
 - Implemented :meth:`sheraf.models.indexation.BaseIndexedModel.__str__`,
+- Implemented :meth:`sheraf.models.indexation.BaseIndexedModel.on_creation` and
+  :meth:`sheraf.models.indexation.BaseIndexedModel.on_deletion` callbacks.
 
 [0.5.23] - 2021-12-22
 =====================